+++ CHANGELOG.md
@@ -1,4 +1,11 @@
 #Change log
+## 1.0-SNAPSHOT
+### Added
+* Add `winners` and `winner` properties to `Ballot`
+
+## Fixed
+* The method `Ballot.hasDuplicates` result was the inversed
+
 ## 1.0-rc1
 ### Added
 * Condorcet method